修改初始化方式，添加mubbostarter实现
